dagster/deployments/fix_pipeline/requirements.txt
pandas
psycopg2-binary
pyarrow 
boto3
dagster
dagster-pandas
dagster-postgres
dagit
dbt-postgres
dbt-core
dbt-duckdb
dagster-webserver
minio


